+++ code/src/HISTORY.md
@@ -1,5 +1,14 @@
 
 
+**1.0.2**
+- Introduction of `ConfigHeader` attribute.
+- Method `CreateDefaultHeader` marked as deprecated.
+- Method `CreateDefaultHeader` replaced by method `CreateExtendedHeader`.
+- Review of the source code documentation as well as of the online help.
+- Version number increased.
+- Package release on `nuget.org`.
+- Wiki and release update on _GitHub_.
+
 **1.0.1**
 - Package migration from .NET Framework 4.7.2 to .NET Standard 2.0 (**BREAKING CHANGES**).
 - Reduction of version number to three digits (**BREAKING CHANGES**).